Seperating methods for files, coderefactor
